requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.4.0
django-decouple==2.1
django-mysql==3.8.1
djangorestframework==3.11.1
mysql==0.0.2
mysqlclient==2.0.1
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
